refactor(sdk): moved the viemLogFromGetEventLogsResponse to the rpc folder
