+++ typescript/sdk/src/block-explorer/etherscan.ts
@@ -1,5 +1,7 @@
 import { Address, HexString } from '@hyperlane-xyz/utils';
 
+import { GetEventLogsResponse } from '../rpc/evm/types.js';
+
 interface EtherscanLikeAPIOptions {
   // Explorers like Blockscout don't require an API key for requests
   apiKey?: string;
@@ -106,7 +108,7 @@ interface GetEventLogs extends BaseEtherscanLikeAPIParams<'logs', 'getLogs'> {
   topic0: string;
 }
 
-export type GetEventLogsResponse = {
+type RawEtherscanGetEventLogsResponse = {
   address: Address;
   blockNumber: HexString;
   data: HexString;
@@ -136,5 +138,18 @@ export async function getLogsFromEtherscanLikeExplorerAPI(
 
   const response = await fetch(requestUrl);
 
-  return handleEtherscanResponse(response);
+  const rawLogs: RawEtherscanGetEventLogsResponse[] =
+    await handleEtherscanResponse(response);
+
+  return rawLogs.map(
+    (rawLogs): GetEventLogsResponse => ({
+      address: rawLogs.address,
+      blockNumber: Number(rawLogs.blockNumber),
+      data: rawLogs.data,
+      logIndex: Number(rawLogs.logIndex),
+      topics: rawLogs.topics,
+      transactionHash: rawLogs.transactionHash,
+      transactionIndex: Number(rawLogs.transactionIndex),
+    }),
+  );
 }
