+++ typescript/sdk/src/block-explorer/utils.ts
@@ -1,13 +1,7 @@
-import { Hex, Log } from 'viem';
-
-import { MultiProvider } from '../index.js';
 import {
   BlockExplorer,
   ExplorerFamily,
 } from '../metadata/chainMetadataTypes.js';
-import { ChainNameOrId } from '../types.js';
-
-import { GetEventLogsResponse } from './etherscan.js';
 
 export function isEvmBlockExplorerAndNotEtherscan(
   blockExplorer: BlockExplorer,
@@ -26,20 +20,4 @@ export function isEvmBlockExplorerAndNotEtherscan(
   };
 
   return byFamily[blockExplorer.family] ?? false;
-}
-
-export function viemLogFromGetEventLogsResponse(
-  log: GetEventLogsResponse,
-): Log {
-  return {
-    address: log.address as Hex,
-    data: log.data as Hex,
-    blockNumber: BigInt(log.blockNumber),
-    transactionHash: log.transactionHash as Hex,
-    logIndex: Number(log.logIndex),
-    transactionIndex: Number(log.transactionIndex),
-    topics: log.topics as [Hex, ...Hex[]],
-    blockHash: null,
-    removed: false,
-  };
 }
